kafka-capture/consumer/requirements.txt
kafka-python==2.0.2
aws-msk-iam-sasl-signer-python==1.0.1